Restrict the missing-value-placeholder to string columns.
For the other columns, it's more confusing than anything. Missingness is
already embedded into the float NaN (with some more digging for true NA
vs NaN), while for integers/logicals, we'll just hard-code them as
-2^31. In both cases, at the R level, we don't really have a choice
because it's not like we could force them to be anything else if the
placeholder wasn't exactly equal to their corresponding NA value anyway.

Also exported the handlers for string placeholders, primarily for re-use
in alabaster.matrix when saving strings.