Switch to a simplified layout for the readers and savers. (#19)
This eliminates the involvement of schemas, allowing each reader and saver to
handle its own affairs without coordination with a separate metadata schema.
The only mandate now is an OBJECT file that controls loader dispatch; all other
files are left to the discretion of the reader/saver specification.

This enables major simplification on several fronts. There is no need to save
metadata and pass it between methods; we don't need to consider a staging
directory and we can save each object directly to its own directory; and thus
there is no need for many of the special directory-manipulating functions. The
instructions for application/extension developers is also greatly simplified.

The old functions are still kept around for back-compatibility, but we stripped
out the documentation to make it clearer for new developers.