Validate all on-disk representation immediately after saving. (#20)
This uses the latest takane C++ validators to check the contents of the
directories created by saveObject. This enforces consistency in the
representations across all frameworks that use the same validators.
Extension developers are now expected to register their own validation
functions, either directly in R or by adding an appropriate C++ validator.