Updated takane  validators and accompanying R bindings.
This adds validation support for various genomic ranges classes, such as the
genomic_ranges itself, sequence_information and various compressed lists.

We also transition to saving factor codes as unsigned integers. This is, on
the whole, more sensible, as downstream consumers don't have to check whether
the codes are negative (which would make no sense outside of missing
placeholders, and we don't need half the integer space for that).