Silently no-op if a type already has a registered reading function.
This allows extension packages to continue to register readers in their own
.onLoad, even if those readers have been added to the default registry. So,
users don't get flooded with lots of irrelevant warning messages because an
extension was deemed important enough to warrant inclusion in the defaults.

This is helpful as extensions would need to deregister readers when they get
unloaded, at which point they also need to re-register on reload. That would
only be possible by calling the registration function in the .onLoad.