Distinguish between scalars and length-1 vectors when saving/loading lists. (#25)
This effectively unboxes all length-1 vectors in a list, by default; this is
probably the more reasonable expectation for other languages that have a
concept of scalars. Users can override this by calling I() on elements that
they want to keep as length-1 vectors, in the same manner as jsonlite.