Added utilities for deduplicating objects within saveObject.
Most of this was migrated was alabaster.matrix and generalized for use outside
of DelayedArrays. The idea is that a saveObject method developer can check if
their object was already saved during a user-defined "deduplication session".
If so, they can just copy/link/symlink from the existing on-disk representation
instead of re-saving the object, reducing compute time and storage costs.

Also added a utility to get an absolute path without requiring normalization,
as the latter could resolve symlinks that a user might want to preserve.