Support ritsuko's custom VLS arrays in character vectors, DFs and lists. (#32)
This allows us to efficiently store variable length strings with compression in
HDF5. The saving/reading code is quite grotesque as Rhdf5lib doesn't
interoperate with the rhdf5 handles, presumably because they're two separate
instances of the same library that don't share their handle records. This
means we need to close all the latter before writing the VLS heap and pointers
with the former, and then re-open the latter for the rest of the function.